+++ lib/kitchen/directions/bake_notes/bake_note_subtitle.rb
@@ -10,7 +10,11 @@ module Kitchen
 
         # Store label information
         note_label = title.children
-        note.target_label(custom_content: note_label.to_s, cases: cases)
+        if cases
+          note.target_label(label_text: 'note', custom_content: note_label.to_s, cases: cases)
+        else
+          note.target_label(custom_content: note_label.to_s)
+        end
 
         title.name = 'h4'
         title.add_class('os-subtitle')