+++ lib/kitchen/directions/bake_notes/bake_numbered_notes/v1.rb
@@ -3,10 +3,10 @@
 module Kitchen::Directions
   module BakeNumberedNotes
     class V1
-      def bake(book:, classes:)
+      def bake(book:, classes:, cases: false)
         classes.each do |klass|
           book.chapters.notes("$.#{klass}").each do |note|
-            bake_note(note: note)
+            bake_note(note: note, cases: cases)
             note.exercises.each do |exercise|
               BakeNumberedNotes.bake_note_exercise(note: note, exercise: exercise)
             end
@@ -17,7 +17,7 @@ module Kitchen::Directions
         end
       end
 
-      def bake_note(note:)
+      def bake_note(note:, cases: false)
         note.wrap_children(class: 'os-note-body')
 
         chapter_count = note.ancestor(:chapter).count_in(:book)
@@ -34,7 +34,7 @@ module Kitchen::Directions
 
         return unless note['use-subtitle']
 
-        BakeNoteSubtitle.v1(note: note)
+        BakeNoteSubtitle.v1(note: note, cases: cases)
       end
     end
   end