+++ spec/directions/bake_notes/bake_numbered_notes/main_spec.rb
@@ -5,8 +5,8 @@ require 'spec_helper'
 RSpec.describe Kitchen::Directions::BakeNumberedNotes do
   it 'calls v1' do
     expect_any_instance_of(Kitchen::Directions::BakeNumberedNotes::V1).to receive(:bake)
-      .with(book: 'book1', classes: 'class1')
-    described_class.v1(book: 'book1', classes: 'class1')
+      .with(book: 'book1', classes: 'class1', cases: false)
+    described_class.v1(book: 'book1', classes: 'class1', cases: false)
   end
 
   it 'calls v2' do